ivp_surface_manager: x86_64 Fix compiler warnings about type truncation

+++ ivp_surface_manager/ivp_compact_grid.hxx
@@ -52,7 +52,7 @@ public:
 
     IVP_FLOAT		radius;
 
-    int			byte_size;
+    intp			byte_size;
     
     IVP_FLOAT		inv_grid_size;
